fix : add the css
