+++ src/app/Assignment1/Question1/Q1.css.jsx
@@ -1,0 +1,9 @@
+const headingStyle = {
+    fontSize: '2rem',
+    color: 'white',
+    textAlign: 'center',
+    marginTop: '50px',
+    fontFamily: 'Segoe UI, Tahoma, Geneva, Verdana, sans-serif',
+  };
+
+  export {headingStyle}
