+++ src/app/Assignment1/Question1/Q1.jsx
@@ -1,12 +1,6 @@
+import { headingStyle } from "./Q1.css";
 function Greeting() {
-  const headingStyle = {
-    fontSize: '2rem',
-    color: 'white',
-    textAlign: 'center',
-    marginTop: '50px',
-    fontFamily: 'Segoe UI, Tahoma, Geneva, Verdana, sans-serif',
-  };
-
+  
   return (
     <h1 style={headingStyle}>Hello, Next.js!</h1>
   );
