+++ src/app/Assignment1/Question3/Q3.css.jsx
@@ -1,0 +1,22 @@
+const appStyle = {
+  padding: '40px',
+  fontFamily: 'Segoe UI, sans-serif',
+  textAlign: 'center',
+};
+
+const selectStyle = {
+  padding: '10px',
+  fontSize: '1rem',
+  marginBottom: '30px',
+};
+
+const textStyle = {
+  fontSize: '1.2rem',
+  fontWeight: 'bold',
+  textAlign: 'center',
+  color: 'white',
+  marginTop: '20px',
+  fontFamily: 'Arial, sans-serif',
+};
+
+export { appStyle, selectStyle, textStyle };
