+++ src/app/Assignment1/Question5/Q5.css.jsx
@@ -1,0 +1,24 @@
+const containerStyle = {
+    fontFamily: 'Segoe UI, sans-serif',
+    padding: '20px',
+    textAlign: 'center',
+  };
+
+  const headingStyle = {
+    fontSize: '1.8rem',
+    marginBottom: '10px',
+    color: '#0070f3',
+  };
+
+  const listStyle = {
+    listStyleType: 'square',
+    padding: '0',
+  };
+
+  const itemStyle = {
+    fontSize: '1.1rem',
+    color: 'white',
+    margin: '8px 0',
+  };
+
+  export{containerStyle,headingStyle,listStyle,itemStyle}
