+++ src/app/Assignment1/Question6/Q6.css.jsx
@@ -1,0 +1,10 @@
+ const buttonStyle = {
+    backgroundColor: color,
+    padding: '10px 20px 10px',
+    borderRadius: '5px',
+    border: 'none',
+    color: 'white'
+
+  };
+
+  export{buttonStyle}
